Remove deprecated logic
