+++ Palace/Reader2/Bookmarks/TPPReadiumBookmark.swift
@@ -11,7 +11,6 @@
 ///
 @objc class TPPBookmarkDictionaryRepresentation: NSObject {
   fileprivate static let annotationIdKey = "annotationId"
-  @objc static let idrefKey = "idref"
   @objc static let hrefKey = "href"
   @objc static let locationKey = "location"
   @objc static let cfiKey = "contentCFI"
@@ -33,18 +32,7 @@
   var page:String?
 
   var location:String
-  var idref:String
   var href:String
-
-  /// The CFI is location information generated from the R1 reader
-  /// which is not usable in R2.
-  ///
-  /// A CFI value refers to the content fragment identifier used to point
-  /// to a specific element within the specified spine item. This was
-  /// consumed by R1, but there has always been very little consistency
-  /// in the values consumed by Library Simplified applications between
-  /// platforms, hence its legacy and optional status.
-  var contentCFI:String?
 
   var progressWithinChapter:Float = 0.0
   var progressWithinBook:Float = 0.0
@@ -61,10 +49,8 @@
   /// Date formatted as per RFC 3339
   let time:String
 
-  /// Deprecated. 
+  /// Deprecated.
   init?(annotationId:String?,
-        contentCFI:String?,
-        idref:String?,
         href:String?,
         chapter:String?,
         page:String?,
@@ -74,30 +60,19 @@
         time:String?,
         device:String?)
   {
-    guard let idref = idref else {
-      Log.error(#file, "Bookmark creation failed init due to nil `idref`.")
-      return nil
-    }
-    
+
     guard let href = href else {
       Log.error(#file, "Bookmark creation failed init due to nil `href`.")
       return nil
     }
 
     self.annotationId = annotationId
-    self.contentCFI = contentCFI
-    self.idref = idref
     self.href = href
     self.chapter = chapter ?? ""
     self.page = page ?? ""
 
-    // TODO: SIMPLY-3655 refactor per spec
-    // This location structure originally comes from R1 Reader's Javascript
-    // and its not available in R2, we are mimicking the structure
-    // in order to pass the needed information to the server
-//    self.location = "{\"href\":\"\(href)\",\"contentCFI\":\"\(contentCFI ?? "")\",\"progressWithinChapter\":\"\(progressWithinChapter)\",\"@type\":\"LocatorHrefProgression\"}"
-    self.location = TPPBookLocation(href: href, type: "LocatorHrefProgression", chapterProgression: progressWithinChapter, totalProgression: progressWithinBook, title: nil, position: nil)?.locationString ?? ""
-
+    self.location = TPPBookLocation(href: href, type: "LocatorHrefProgression", chapterProgression: progressWithinChapter, totalProgression: progressWithinBook, title: chapter, position: nil)?.locationString ?? ""
+    
     self.progressWithinChapter = progressWithinChapter
     self.progressWithinBook = progressWithinBook
     self.time = time ?? NSDate().rfc3339String()
@@ -106,9 +81,7 @@
 
   init?(dictionary:NSDictionary)
   {
-    guard let contentCFI = dictionary[TPPBookmarkDictionaryRepresentation.cfiKey] as? String,
-      let idref = dictionary[TPPBookmarkDictionaryRepresentation.idrefKey] as? String,
-      let href = dictionary[TPPBookmarkDictionaryRepresentation.hrefKey] as? String,
+    guard let href = dictionary[TPPBookmarkDictionaryRepresentation.hrefKey] as? String,
       let location = dictionary[TPPBookmarkDictionaryRepresentation.locationKey] as? String,
       let time = dictionary[TPPBookmarkDictionaryRepresentation.timeKey] as? String else {
         Log.error(#file, "Bookmark failed to init from dictionary.")
@@ -120,8 +93,6 @@
     } else {
       self.annotationId = nil
     }
-    self.contentCFI = contentCFI
-    self.idref = idref
     self.href = href
     self.location = location
     self.time = time
@@ -139,8 +110,6 @@
   var dictionaryRepresentation:NSDictionary {
     return [
       TPPBookmarkDictionaryRepresentation.annotationIdKey: self.annotationId ?? "",
-      TPPBookmarkDictionaryRepresentation.cfiKey: self.contentCFI ?? "",
-      TPPBookmarkDictionaryRepresentation.idrefKey: self.idref,
       TPPBookmarkDictionaryRepresentation.hrefKey: self.href,
       TPPBookmarkDictionaryRepresentation.chapterKey: self.chapter ?? "",
       TPPBookmarkDictionaryRepresentation.pageKey: self.page ?? "",
@@ -151,28 +120,16 @@
       TPPBookmarkDictionaryRepresentation.bookProgressKey: self.progressWithinBook
     ]
   }
-  
+
   override func isEqual(_ object: Any?) -> Bool {
     guard let other = object as? TPPReadiumBookmark else {
       return false
     }
 
-    if let contentCFI = self.contentCFI,
-      let otherContentCFI = other.contentCFI,
-      contentCFI.count > 0 && otherContentCFI.count > 0 {
-      // R1
-      return self.idref == other.idref
-        && self.contentCFI == other.contentCFI
-        && self.location == other.location
-        && self.chapter == other.chapter
-    } else {
-      // R2
-      return self.idref == other.idref
-        && self.href == other.href
-        && self.progressWithinBook =~= other.progressWithinBook
-        && self.progressWithinChapter =~= other.progressWithinChapter
-        && self.chapter == other.chapter
-    }
+    return self.href == other.href
+      && self.progressWithinBook =~= other.progressWithinBook
+      && self.progressWithinChapter =~= other.progressWithinChapter
+      && self.chapter == other.chapter
   }
 }
